Continued development

+++ src/ABC_Util.h
@@ -94,36 +94,36 @@ extern "C" {
                                             }
 #define ABC_BUF_PTR(x)                      ((x).p)
 #define ABC_BUF_REALLOC(buf, count)         { \
-                                                int __abc_buf_realloc_count__ = (sizeof(*((buf).p)) * count); \
+                                                unsigned long __abc_buf_realloc_count__ = (sizeof(*((buf).p)) * count); \
                                                 (buf).p = realloc((buf).p, __abc_buf_realloc_count__); \
                                                 (buf).end = (buf).p +      __abc_buf_realloc_count__; \
                                             }
 #define ABC_BUF_EXPAND(buf, count)          { \
-                                                int __abc_buf_expand_count__ = (((buf).end)-((buf).p)) + (sizeof(*((buf).p)) * count); \
+                                                unsigned long __abc_buf_expand_count__ = (((buf).end)-((buf).p)) + (sizeof(*((buf).p)) * count); \
                                                 (buf).p = realloc((buf).p, __abc_buf_expand_count__); \
                                                 (buf).end = (buf).p +      __abc_buf_expand_count__; \
                                             }
 #define ABC_BUF_APPEND_PTR(buf, ptr, count) { \
-                                                int __abc_buf_append_resize__ = (int) (((buf).end)-((buf).p)) + (sizeof(*((buf).p)) * count); \
+                                                unsigned long __abc_buf_append_resize__ = (unsigned long) (((buf).end)-((buf).p)) + (sizeof(*((buf).p)) * count); \
                                                 (buf).p = realloc((buf).p, __abc_buf_append_resize__); \
                                                 (buf).end = (buf).p +      __abc_buf_append_resize__; \
                                                 memcpy((buf).end - count, ptr, count); \
                                             }
 #define ABC_BUF_APPEND(dst, src)            { \
-                                                int __abc_buf_append_size__ = (((src).end)-((src).p)); \
-                                                int __abc_buf_append_resize__ = (((dst).end)-((dst).p)) + __abc_buf_append_size__ ; \
+                                                unsigned long __abc_buf_append_size__ = (unsigned long) (((src).end)-((src).p)); \
+                                                unsigned long __abc_buf_append_resize__ = (unsigned long) (((dst).end)-((dst).p)) + __abc_buf_append_size__ ; \
                                                 (dst).p = realloc((dst).p, __abc_buf_append_resize__); \
                                                 (dst).end = (dst).p +      __abc_buf_append_resize__; \
                                                 memcpy((dst).end - __abc_buf_append_size__, (src).p,  __abc_buf_append_size__); \
                                             }
 #define ABC_BUF_DUP(dst, src)               { \
-                                                int __abc_buf_dup_size__ = (int) (((src).end)-((src).p)); \
+                                                unsigned long __abc_buf_dup_size__ = (int) (((src).end)-((src).p)); \
                                                 (dst).p = malloc(__abc_buf_dup_size__); \
                                                 (dst).end = (dst).p + __abc_buf_dup_size__; \
                                                 memcpy((dst).p, (src).p, __abc_buf_dup_size__); \
                                             }
 #define ABC_BUF_DUP_PTR(buf, ptr, size)     { \
-                                                int __abc_buf_dup_size__ = (int) size; \
+                                                unsigned long __abc_buf_dup_size__ = (int) size; \
                                                 (buf).p = malloc(__abc_buf_dup_size__); \
                                                 (buf).end = (buf).p + __abc_buf_dup_size__; \
                                                 memcpy((buf).p, ptr, __abc_buf_dup_size__); \
